refactor: 글로벌 패키지 생성
